requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
PyYAML==5.4.1
schedule==1.1.0
selenium==3.141.0
soupsieve==2.2.1
urllib3==1.26.4
